+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 1.1.0.0
 
+* Add `HasName`
+* Remove `HasComment` instance from `Named`
 * Add `bool` to `Huddle` module
 * Removed most `Show` instances from `Huddle` as they were unlawful
 * Added `ctrTerm` and `ctrResult` field accessors to `CBORTermResult`